refactor(tokens): split monolithic design tokens into modular files
Improve organization and maintainability by splitting the design tokens package into separate modular files for colors, spacing, typography, borders, shadows, layout, animation, and themes. Update imports in the Crisis Unleashed app and improve type safety with `as const` assertions. Fix FactionId type in favicon renderer.
