+++ packages/tokens/src/borders.ts
@@ -1,0 +1,18 @@
+/**
+ * Border tokens
+ */
+
+/**
+ * Border radius
+ */
+export const borderRadius = {
+  none: '0',
+  sm: '0.125rem',    // 2px
+  default: '0.25rem', // 4px
+  md: '0.375rem',     // 6px
+  lg: '0.5rem',       // 8px
+  xl: '0.75rem',      // 12px
+  '2xl': '1rem',      // 16px
+  '3xl': '1.5rem',    // 24px
+  full: '9999px',
+} as const;
