+++ packages/tokens/src/colors.ts
@@ -1,0 +1,77 @@
+/**
+ * Color palette tokens
+ */
+export const colors = {
+  // Primary colors
+  primary: {
+    50: '#e3f2fd',
+    100: '#bbdefb',
+    200: '#90caf9',
+    300: '#64b5f6',
+    400: '#42a5f5',
+    500: '#2196f3',
+    600: '#1e88e5',
+    700: '#1976d2',
+    800: '#1565c0',
+    900: '#0d47a1',
+  },
+  
+  // Secondary colors
+  secondary: {
+    50: '#f3e5f5',
+    100: '#e1bee7',
+    200: '#ce93d8',
+    300: '#ba68c8',
+    400: '#ab47bc',
+    500: '#9c27b0',
+    600: '#8e24aa',
+    700: '#7b1fa2',
+    800: '#6a1b9a',
+    900: '#4a148c',
+  },
+  
+  // Accent colors
+  accent: {
+    50: '#fff3e0',
+    100: '#ffe0b2',
+    200: '#ffcc80',
+    300: '#ffb74d',
+    400: '#ffa726',
+    500: '#ff9800',
+    600: '#fb8c00',
+    700: '#f57c00',
+    800: '#ef6c00',
+    900: '#e65100',
+  },
+  
+  // Neutral colors
+  neutral: {
+    50: '#fafafa',
+    100: '#f5f5f5',
+    200: '#eeeeee',
+    300: '#e0e0e0',
+    400: '#bdbdbd',
+    500: '#9e9e9e',
+    600: '#757575',
+    700: '#616161',
+    800: '#424242',
+    900: '#212121',
+  },
+  
+  // Semantic colors
+  success: '#4caf50',
+  warning: '#ff9800',
+  error: '#f44336',
+  info: '#2196f3',
+  
+  // Game-specific colors
+  health: '#ff5252',
+  mana: '#2979ff',
+  experience: '#ffeb3b',
+  gold: '#ffc107',
+  
+  // Common colors
+  black: '#000000',
+  white: '#ffffff',
+  transparent: 'transparent',
+} as const;
