+++ packages/tokens/src/typography.ts
@@ -1,0 +1,60 @@
+/**
+ * Typography tokens: font families, sizes, weights, and line heights
+ */
+
+/**
+ * Font families
+ */
+export const fontFamily = {
+  sans: 'Inter, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif',
+  serif: 'Georgia, Cambria, "Times New Roman", Times, serif',
+  mono: 'Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace',
+  game: '"Press Start 2P", cursive',
+  display: '"Audiowide", cursive',
+} as const;
+
+/**
+ * Font sizes
+ */
+export const fontSize = {
+  xs: '0.75rem',     // 12px
+  sm: '0.875rem',    // 14px
+  base: '1rem',      // 16px
+  lg: '1.125rem',    // 18px
+  xl: '1.25rem',     // 20px
+  '2xl': '1.5rem',   // 24px
+  '3xl': '1.875rem', // 30px
+  '4xl': '2.25rem',  // 36px
+  '5xl': '3rem',     // 48px
+  '6xl': '3.75rem',  // 60px
+  '7xl': '4.5rem',   // 72px
+  '8xl': '6rem',     // 96px
+  '9xl': '8rem',     // 128px
+} as const;
+
+/**
+ * Font weights
+ */
+export const fontWeight = {
+  thin: '100',
+  extralight: '200',
+  light: '300',
+  normal: '400',
+  medium: '500',
+  semibold: '600',
+  bold: '700',
+  extrabold: '800',
+  black: '900',
+} as const;
+
+/**
+ * Line heights
+ */
+export const lineHeight = {
+  none: '1',
+  tight: '1.25',
+  snug: '1.375',
+  normal: '1.5',
+  relaxed: '1.625',
+  loose: '2',
+} as const;
